refactor(bot): encapsulate the speeches on the bot module #22
