+++ .env.example
@@ -1,6 +1,7 @@
 GRAPHQL_URL='http://localhost:3003/graphql'
 APP_DOMAIN='http://localhost:5000'
 API_URL='http://localhost:3000'
+SPEECH_VERSION='v0'
 
 # BONDE's credentials to make authenticated requests with GraphQL
 SERVER_AUTH_EMAIL='email'
