+++ bot/helpers/index.js
@@ -1,1 +1,2 @@
 export { default as receive } from './receive'
+export { default as replyWithSave } from './reply-with-save'
