+++ bot/helpers/reply-with-save.js
@@ -1,0 +1,13 @@
+import * as botInteractions from '../interactions'
+
+//
+// Reply function interface to save the bot's reply text
+// before send it to user.
+//
+export default ({ botData, payload, originalReply }) => (message, action) => {
+  const interaction = { is_bot: true, message, action }
+
+  botInteractions.save({ botData, payload, interaction })
+    .then(data => { originalReply(message); return data })
+    .catch(error => console.error(`${error}`.red))
+}
