+++ bot/speeches/index.js
@@ -1,0 +1,1 @@
+export { default as actionStrategy } from './action-strategy'
