+++ bot/speeches/utils/index.js
