+++ bot/speeches/utils/reply-text.js
