+++ bot/speeches/v0/index.js
@@ -1,0 +1,1 @@
+export { default as speech } from './speech'
